socket/item.c, socket/loop.c, socket/lowlevel.c, socket/metaserver.c: Fix some   format string bugs (found using the GCC format __attribute__ for some   crossfire functions).
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@11549 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ socket/item.c
@@ -775,7 +775,7 @@ void esrv_move_object(object *pl, tag_t to, tag_t tag, long nrof) {
 
     op = esrv_get_ob_from_count(pl, tag);
     if (!op) {
-        LOG(llevDebug, "Player '%s' tried to move an unknown object (%ld)\n", pl->name, tag);
+        LOG(llevDebug, "Player '%s' tried to move an unknown object (%lu)\n", pl->name, (unsigned long)tag);
         return;
     }
